Pull createUserContextGenerator into own utility function
